requirements.txt
Flask
Flask-Dance
Flask-Login
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
cryptography
gunicorn
pyodbc 

